requirements.txt
cadquery
jupyterlab
https://github.com/shimwell/wheels/raw/refs/heads/main/moab/moab-wheels-ubuntu-latest/moab-5.5.1-cp311-cp311-manylinux_2_28_x86_64.whl
https://github.com/shimwell/wheels/raw/refs/heads/main/openmc/openmc-0.15.3-cp311-cp311-manylinux_2_28_x86_64.whl
paramak
remove_dagmc_tags
pyvista
pytest
nbformat
nbconvert
ipykernel
jupyter-book
openmc_source_plotter
openmc_regular_mesh_plotter
openmc_cylindrical_mesh_plotter
openmc_depletion_plotter
openmc_data_downloader
openmc_data
dagmc_geometry_slice_plotter
cad_to_dagmc
openmc-plasma-source
neutronics_material_maker[density]
git+https://github.com/svalinn/pydagmc
